# api/openapi.yaml
openapi: "3.0.3"
info:
  title: bodega API
  description: REST API for the bodega package repository manager.
  version: 0.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/packages/{type}/{name}:
    get:
      summary: Get a single package
      operationId: getPackage
      tags: [packages]
      parameters:
        - $ref: "#/components/parameters/PackageType"
        - $ref: "#/components/parameters/PackageName"
      responses:
        "200":
          description: Package manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageManifest"
        "404":
          description: Package not found
    delete:
      summary: Delete a package entry
      description: |
        Deletes a package manifest entry. Frozen packages cannot be deleted.
        Same access control as POST.
      operationId: deletePackage
      tags: [packages]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PackageType"
        - $ref: "#/components/parameters/PackageName"
      responses:
        "200":
          description: Package deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: deleted
                  type:
                    type: string
                  name:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Package not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API token generated with `bodega token generate`. Required for
        POST and DELETE when the server accepts connections from non-localhost IPs.
  parameters:
    PackageType:
      name: type
      in: path
      required: true
      schema:
        type: string
        enum: [apt, git, pypi, binary, gomod, helm, npm, cargo]
    PackageName:
      name: name
      in: path
      required: true
      schema:
        type: string
  responses:
...
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        text/plain:
          schema:
            type: string
    Forbidden:
      description: IP not in admin_permit_cidr or entry is frozen
      content:
        text/plain:
          schema:
            type: string
  schemas:
...
    PackageManifest:
      type: object
      required: [name, type]
      properties:
        config_version:
          type: integer
          description: Schema version
        name:
          type: string
          description: Package name
        type:
          type: string
          enum: [apt, git, pypi, binary, gomod, helm, npm]
        description:
          type: string
        dep_policy:
          type: string
          enum: [none, direct, transitive]
        versions:
          type: array
          items:
            $ref: "#/components/schemas/VersionEntry"
    VersionEntry:
      type: object
      properties:
        version:
          type: string
        url:
          type: string
        mode:
          type: string
          enum: [hosted, proxy]
        version_constraint:
          type: string
          enum: [exact, compatible, patch, any]
        ref:
          type: string
          description: Git ref (tag, branch, SHA)
        source:
          type: string
          enum: [release, clone]
          description: Git source mode
        source_name:
          type: string
          description: Upstream apt package name
        build_cmd:
          type: string
        deb_glob:
          type: string
        filename:
          type: string
          description: Binary filename override
        sha256:
          type: string
          description: Expected SHA-256 for binary downloads
        app_version:
          type: string
          description: Helm app version
        required_by:
          type: array
          items:
            type: string
        checksum:
          $ref: "#/components/schemas/Checksum"
        checksum_verified:
          type: boolean
        platform:
          type: string
          description: Target platform (e.g. linux/amd64)
        artifact_size:
          type: integer
          format: int64
          description: Size in bytes
        hidden:
          type: boolean
        frozen:
          type: boolean
        description:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    Checksum:
      type: object
      properties:
        algorithm:
          type: string
          enum: [sha256, sha1, md5]
        value:
          type: string
          description: Hex-encoded digest
...